+++ spec/generators/twirp_rspec_generator_spec.rb
@@ -2,11 +2,11 @@ require 'spec_helper'
 require 'generator_spec'
 
 RSpec.describe Twirp::RspecGenerator, type: :generator do
-  destination File.expand_path('../tmp/dummy', __dir__)
+  destination File.expand_path('../../tmp/dummy', __dir__)
 
   before(:all) do
     prepare_destination
-    FileUtils.cp_r File.expand_path('dummy', __dir__), File.expand_path('../tmp', __dir__)
+    FileUtils.cp_r File.expand_path('../dummy', __dir__), File.expand_path('../../tmp', __dir__)
     run_generator
   end